+++ src/CognitiveServices/CognitiveServices/CognitiveServicesAccount/NewAzureCognitiveServicesAccount.cs
@@ -245,6 +245,24 @@ namespace Microsoft.Azure.Commands.Management.CognitiveServices
                             }
                         }
                     }
+
+                    if (Type.Equals("Face", StringComparison.InvariantCultureIgnoreCase) || Type.Equals("CognitiveServices", StringComparison.InvariantCultureIgnoreCase))
+                    {
+                        if (Force.IsPresent)
+                        {
+                            WriteWarning(Resources.NewAccount_LegalTerm_NotPolice);
+                        }
+                        else
+                        {
+                            bool yesToAll = false, noToAll = false;
+                            if (!ShouldContinue(Resources.NewAccount_LegalTerm_NotPolice, "Notice", true, ref yesToAll, ref noToAll))
+                            {
+                                return;
+                            }
+                        }
+                    }
+
+
                     try
                     {
                         CognitiveServicesAccount createAccountResponse = CognitiveServicesClient.Accounts.Create(