+++ src/Databricks/exports/Update-AzDatabricksWorkspace.ps1
@@ -46,6 +46,7 @@ To create the parameters described below, construct a hash table containing the 
 
 INPUTOBJECT <IDatabricksIdentity>: Identity parameter.
   [Id <String>]: Resource identity path
+  [PeeringName <String>]: The name of the workspace vNet peering.
   [ResourceGroupName <String>]: The name of the resource group. The name is case insensitive.
   [SubscriptionId <String>]: The ID of the target subscription.
   [WorkspaceName <String>]: The name of the workspace.