+++ src/Storage/Storage.Management/Blob/GetAzureStorageBlobServiceProperties.cs
@@ -22,10 +22,12 @@ namespace Microsoft.Azure.Commands.Management.Storage
     using System.Management.Automation;
     using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
     using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
+    using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
     /// <summary>
     /// Modify Azure Storage service properties
     /// </summary>
+    [CmdletOutputBreakingChange(typeof(PSBlobServiceProperties), ChangeDescription = "The deprecated proeprty RestorePolicy.LastEnabledTime will be removed in a future release.")]
     [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + StorageBlobServiceProperty, DefaultParameterSetName = AccountNameParameterSet), OutputType(typeof(PSBlobServiceProperties))]
     public class GetAzStorageBlobServicePropertyCommand : StorageBlobBaseCmdlet
     {