+++ src/FrontDoor/FrontDoor/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.6.1
+* Fixed an issue where an exception is being thrown when Enum.Parse tries to coerce a null value to an Enabled or Disabled enum values [#12344]
+
 ## Version 1.6.0
 * Updated module to use API 2020-05-01
 * Added Private link support for Storage, Keyvault and Web App Service resources