+++ src/Sql/Sql/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added BackupStorageRedundancy to `New-AzSqlInstance` and `Get-AzSqlInstance`
+
+## Version 2.9.1
 * Fixed potential server name case insensitive error in `New-AzSqlServer` and `Set-AzSqlServer`
 * Fixed wrong database name returned on existing database error in `New-AzSqlDatabaseSecondary`
 * Added operation parameters and steps in `Get-AzSqlInstanceOperation`